✨ E94KR6 website: promote home preview to root
