--- website/src/pages/index.module.css
@@ -1,126 +1,0 @@
-.placeholderPage {
-  display: grid;
-  min-height: calc(100vh - 120px);
-  place-items: center;
-  padding: 3rem 1rem 5rem;
-}
-
-.heroSheet {
-  border: 1px solid rgb(var(--border));
-  display: grid;
-  gap: 1rem;
-  justify-items: start;
-  max-width: 44rem;
-  padding: 2.2rem 2rem 1.9rem;
-  position: relative;
-  width: 100%;
-}
-
-.heroSheet::before,
-.heroSheet::after {
-  border-top: 1px solid rgb(var(--faint));
-  content: "";
-  height: 0;
-  position: absolute;
-  width: 18px;
-}
-
-.heroSheet::before {
-  left: 10px;
-  top: 10px;
-}
-
-.heroSheet::after {
-  bottom: 10px;
-  right: 10px;
-}
-
-.logoIcon {
-  height: 4rem;
-  margin: 0;
-  width: 4rem;
-}
-
-.wordmark {
-  height: auto;
-  max-width: 24rem;
-  width: min(100%, 24rem);
-}
-
-.pixelTag {
-  color: rgb(var(--faint));
-  font-family: var(--font-pixel), var(--ifm-font-family-monospace);
-  font-size: 0.68rem;
-  letter-spacing: 0.1em;
-}
-
-.subtitle {
-  color: rgb(var(--muted));
-  margin: 0;
-  max-width: 58ch;
-}
-
-.docsLink {
-  border: 1px solid rgb(var(--fg));
-  color: rgb(var(--fg));
-  display: inline-block;
-  font-size: 0.9rem;
-  padding: 0.56rem 0.94rem;
-  text-decoration: none;
-  transition:
-    border-color 180ms ease,
-    color 180ms ease;
-}
-
-.docsLink:focus-visible {
-  outline: 2px solid rgb(var(--fg));
-  outline-offset: 2px;
-}
-
-.docsLink:hover {
-  border-color: rgb(var(--accent));
-  color: rgb(var(--accent));
-  text-decoration: underline;
-}
-
-.artifact {
-  border: 1px solid rgb(var(--border));
-  color: rgb(var(--fg));
-  font-family: var(--ifm-font-family-monospace);
-  font-size: clamp(1.25rem, 4vw, 2.2rem);
-  font-weight: 700;
-  letter-spacing: 0.02em;
-  line-height: 1.1;
-  margin: 0;
-  padding: 0.8rem 0.95rem;
-  white-space: nowrap;
-  width: fit-content;
-}
-
-.marginNote {
-  border-left: 1px solid rgb(var(--border));
-  color: rgb(var(--faint));
-  font-family: var(--ifm-font-family-monospace);
-  font-size: 0.76rem;
-  letter-spacing: 0.02em;
-  line-height: 1.45;
-  margin: 0.2rem 0 0;
-  max-width: 60ch;
-  padding-left: 0.72rem;
-}
-
-@media (max-width: 768px) {
-  .heroSheet {
-    gap: 0.84rem;
-    padding: 1.4rem 1.2rem;
-  }
-
-  .artifact {
-    max-width: 100%;
-    white-space: normal;
-  }
-
-  .wordmark {
-    max-width: 18rem;
-  }
-}
